requirements.txt
numpy
pillow
pycairo
ffmpeg-python